working on delete order view delete / before $ in urls

+++ store/views.py
@@ -321,11 +321,14 @@ def edit_order(request, id):
 
 @login_required(login_url='login')
 def delete_order(request, id):
-    delOrder = get_object_or_404(Order, pk=id).delete()
-    delOrder.delete()
-    messages.success(request,'Order #{0} was successfully deleted.'.format(id))
+    delOrder = get_object_or_404(Order, pk=id)
+    if delOrder:
+        delOrder.delete()
+        messages.success(request,'Order #{0} was successfully deleted.'.format(id))
+        return HttpResponseRedirect('/order/view')
+    else:
+        messages.error(request, 'Could not delete order #{}. Order does not exist.'.format(id))
     return HttpResponseRedirect(reverse('store.views.view_order'))
-
     # try:
     #     delOrder = Order.objects.get(pk=id)
     # except Order.DoesNotExist:  # expression as identifier:
